requirements.txt
flask==2.3.2
twilio==8.13.0
requests==2.31.0
pydub==0.25.1
python-dotenv==1.0.0